refactor: improve rendering speed for large task lists
