+++ src/components/TaskListItem.tsx
@@ -1,5 +1,5 @@
 import { Checkbox, ListItemButton, Stack, styled } from "@mui/material";
-import { forwardRef, useRef } from "react";
+import { forwardRef, useMemo, useRef } from "react";
 import { Task } from "../utils/task";
 import TaskBody from "./TaskBody";
 import TaskDates from "./TaskDates";
@@ -34,7 +34,6 @@ interface TaskListItemProps {
 const TaskListItem = forwardRef<HTMLDivElement, TaskListItemProps>(
   (props, ref) => {
     const { task, focused, onClick, onCheckboxClick, onBlur, onFocus } = props;
-
     const checkboxRef = useRef<HTMLButtonElement>(null);
     const menuRef = useRef<HTMLUListElement>(null);
     const menuButtonRef = useRef<HTMLButtonElement>(null);
@@ -59,55 +58,59 @@ const TaskListItem = forwardRef<HTMLDivElement, TaskListItemProps>(
       }
     };
 
-    return (
-      <TaskItemButton
-        ref={ref}
-        aria-label="Task"
-        aria-current={focused}
-        onClick={handleItemClick}
-        onFocus={onFocus}
-        onBlur={onBlur}
-        dense
-      >
-        <Stack
-          px={{ xs: 0.5, sm: 0 }}
-          direction="row"
-          spacing={0.5}
-          sx={{ width: "100%" }}
+    return useMemo(
+      () => (
+        <TaskItemButton
+          ref={ref}
+          aria-label="Task"
+          aria-current={focused}
+          onClick={handleItemClick}
+          onFocus={onFocus}
+          onBlur={onBlur}
+          dense
         >
-          <div>
-            <Checkbox
-              ref={checkboxRef}
-              inputProps={{
-                "aria-label": "Complete task",
-                "aria-checked": task.completed,
-              }}
-              onClick={onCheckboxClick}
-              edge="start"
-              checked={task.completed}
-              tabIndex={-1}
-            />
-          </div>
           <Stack
-            direction="column"
-            style={{
-              paddingTop: 10,
-              paddingBottom: 10,
-              flex: "auto",
-            }}
+            px={{ xs: 0.5, sm: 0 }}
+            direction="row"
+            spacing={0.5}
+            sx={{ width: "100%" }}
           >
-            <TaskBody task={task} />
-            <TaskDates task={task} />
+            <div>
+              <Checkbox
+                ref={checkboxRef}
+                inputProps={{
+                  "aria-label": "Complete task",
+                  "aria-checked": task.completed,
+                }}
+                onClick={onCheckboxClick}
+                edge="start"
+                checked={task.completed}
+                tabIndex={-1}
+              />
+            </div>
+            <Stack
+              direction="column"
+              style={{
+                paddingTop: 10,
+                paddingBottom: 10,
+                flex: "auto",
+              }}
+            >
+              <TaskBody task={task} />
+              <TaskDates task={task} />
+            </Stack>
+            <div>
+              <TaskListItemMenu
+                task={task}
+                menuRef={menuRef}
+                menuButtonRef={menuButtonRef}
+              />
+            </div>
           </Stack>
-          <div>
-            <TaskListItemMenu
-              task={task}
-              menuRef={menuRef}
-              menuButtonRef={menuButtonRef}
-            />
-          </div>
-        </Stack>
-      </TaskItemButton>
+        </TaskItemButton>
+      ),
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+      [task, focused]
     );
   }
 );
